pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nas-streamliner"
version = "0.1.0"
description = "Structured ingest and classification pipeline for NAS-based video workflows."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
